cart/wishlist Process done

+++ productPHP/deleteProductProcess.php
@@ -6,7 +6,7 @@ if(isset($_GET["id"])){
 
     $pid = $_GET["id"];
 
-    Database::iud("DELETE FROM `product` WHERE `id`='".$pid."'");
+    Database::iud("DELETE FROM `product` WHERE `id`='".$pid."' AND `user_id`='".$_SESSION["u"]["id"]."'");
 
 }else{
